requirements.txt

pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
xgboost>=1.7
flask>=2.3
pydantic>=2.0
pyyaml>=6.0
uvicorn>=0.23
matplotlib>=3.7
seaborn>=0.12
